### 2025-02-20 #7824 @editorialbot
[PRE REVIEW]: Stresslog: A python package for modelling wellbore stability in inclined stress states
**Submitting author:** @GeoArkadeep (<a href="http://orcid.org/0009-0008-3209-8125">Arkadeep Ghosh</a>)
**Repository:** https://github.com/GeoArkadeep/Stresslog
**Branch with paper.md** (empty if default branch):
**Version:** v1.6.1
