+++ paper/paper.bib
@@ -5,7 +5,7 @@
     pages = {103580},
     year = {2022},
     issn = {1750-5836},
-    doi = {https://doi.org/10.1016/j.ijggc.2022.103580},
+    doi = {10.1016/j.ijggc.2022.103580},
     url = {https://www.sciencedirect.com/science/article/pii/S1750583622000020},
     author = {Nicholas Thompson and Jamie Stuart Andrews and Long Wu and Renata Meneguolo}
 }
@@ -17,7 +17,7 @@
     publisher={Cambridge University Press},
     author={Flemings, Peter B.},
     year={2021},
-    doi = {https://doi.org/10.1017/9781107326309.010},
+    doi = {10.1017/9781107326309.010},
     pages={211–232}
 }
 
@@ -28,7 +28,7 @@
     pages = {2-11},
     year = {2013},
     issn = {0264-8172},
-    doi = {https://doi.org/10.1016/j.marpetgeo.2013.04.007},
+    doi = {10.1016/j.marpetgeo.2013.04.007},
     url = {https://www.sciencedirect.com/science/article/pii/S0264817213000913},
     author = {Jincai Zhang}
 }
@@ -65,7 +65,7 @@
     number={02},
     pages={68--73},
     year={2001},
-    doi={https://doi.org/10.2118/56017-PA},
+    doi={10.2118/56017-PA},
     publisher={SPE}
 }
 
@@ -75,7 +75,7 @@
     booktitle={SPE Asia Pacific Oil and Gas Conference and Exhibition},
     pages={SPE--54356},
     year={1999},
-    doi={https://doi.org/10.2118/54356-MS},
+    doi={10.2118/54356-MS},
     organization={SPE}
 }
 
@@ -87,7 +87,7 @@
     pages = {86-100},
     year = {2011},
     issn = {0920-4105},
-    doi = {https://doi.org/10.1016/j.petrol.2011.07.014},
+    doi = {10.1016/j.petrol.2011.07.014},
     url = {https://www.sciencedirect.com/science/article/pii/S0920410511001756},
     author = {Kjetil Thorsen}
 }
@@ -101,7 +101,7 @@
     year = {2003},
     note = {Special Issue of the IJRMMS: Rock Stress Estimation ISRM Suggested Methods and Associated Supporting Papers},
     issn = {1365-1609},
-    doi = {https://doi.org/10.1016/j.ijrmms.2003.07.001},
+    doi = {10.1016/j.ijrmms.2003.07.001},
     url = {https://www.sciencedirect.com/science/article/pii/S1365160903001175},
     author = {M.D. Zoback and C.A. Barton and M. Brudy and D.A. Castillo and T. Finkbeiner and B.R. Grollimund and D.B. Moos and P. Peska and C.D. Ward and D.J. Wiprut}
 }
@@ -154,7 +154,7 @@
   number={B4},
   pages={5039--5057},
   year={1992},
-  doi={https://doi.org/10.1029/91JB02175},
+  doi={10.1029/91JB02175},
   publisher={Wiley Online Library}
 }
 
@@ -169,6 +169,6 @@
     month = {06},
     issn = {1064-6671},
     doi = {10.2118/56862-PA},
-    url = {https://doi.org/10.2118/56862-PA},
+    url = {10.2118/56862-PA},
     eprint = {https://onepetro.org/DC/article-pdf/14/02/85/2590516/spe-56862-pa.pdf},
 }
