### 2025-02-20 aae17e1
Shrunk paper a wee bit, fix some dois
