+++ paper/paper.md
@@ -129,7 +129,7 @@ The well data from Equinor Northern Lights dataset has been used as the example 
 
 ![Model of Northern Lights Eos Well showing the Drake I, II and IntraMarine formations.](./figures/WellPlot.png)
 
-![Fracture motifs calculated at 2624.5m and superimposed onto the image log. A perfectly vertical wellbore has been assumed in this example.](./figures/overlay.png)
+![Fracture motifs calculated at 2624.5m and superimposed onto the image log.](./figures/overlay.png)
 
 It is not being suggested that this interpretation of the data is preferred over any other, the analysis by [@Thompson2022] is much more comprehensive.
 
