+++ paper/paper.bib
@@ -41,6 +41,7 @@
     number={B7},
     pages={12791--12811},
     year={1995},
+    doi={10.1029/95jb00319},
     publisher={Wiley Online Library}
 }
 
