### 2025-02-20 #7824 @GeoArkadeep
Thank you @kthyng !
I will work on the doi related issues and try to bring down the paper length to meet the JOSS specifications.

I have gone through the reviewers list and will suggest reviewers tomorrow.
