### 2025-02-24 30ccc16
Trim paper to correct length
