+++ paper/paper.md
@@ -4,7 +4,7 @@ tags:
   - Python
   - Drilling
   - Geomechanics
-  - Pore pressure
+  - pore-pressure
   - Mud loss
   - Wellbore stability
 authors:
@@ -23,19 +23,17 @@ bibliography: paper.bib
 
 # Summary
 
-This package is meant to be used by researchers and practitioners working in the field of geomechanics. It uses a collection of algorithms used to iteratively model the sate of stress underground, given a well log. The computations use full 6 component stress tensor (as calculated using [@pevska1995]) which allows modelling of inclined wellbores in inclined states of stress. Stresslog estimates, among other things, pore pressure (with and without unloading effects) from sonic, resistivity and d-exponent, principal stresses, hoop stresses, compressive strength, tensile strength, the three moduli of elasticity and sanding tendency.
+This package is meant to be used by researchers and practitioners working in the field of geomechanics. It uses a collection of algorithms used to iteratively model the sate of stress underground, given a well log. The computations use 6 component stress tensor (as calculated using [@pevska1995]) which allows modelling of inclined wellbores in inclined states of stress. Stresslog estimates pore-pressure (with and without unloading effects) from sonic, resistivity and d-exponent, principal stresses, hoop stresses, compressive strength, tensile strength, the three moduli of elasticity and sanding tendency.
 
 # Statement of need
 
-Stresslog has been designed to help with pre-drill, post-drill and realtime geomechanical calculations. Stresslog is aimed at empowering researchers with a simple to use and comprehensive 1D mechanical earth modelling tool that is freely available and which researchers can modify to apply their own methods when neccessary, while allowing practitioners to use the pre-defined algorithms to calculate solutions, iteratively process and export well log data.
+Stresslog has been designed to help with pre-drill, post-drill and realtime geomechanical calculations. Stresslog is aimed at empowering researchers with a comprehensive 1D mechanical earth modelling tool that is freely available and which researchers can modify to apply their own methods when neccessary, while allowing practitioners to use industry-standard algorithms to calculate solutions and export well log data.
 
 # Methodology
 
-Overburden gradient [@traugott1997], pore pressure [@Zhang20132; @Flemings2021], minimum horizontal stress [@Daines1982; @zoback1992], rock strength [@lal1999; @horsrud2001] and other parameters are calculated considering the given well-logging, deviation, and formation data. The maximum horizontal stress is estimated by applying stress polygon [@ZOBACK2003] for every depth-sample. Borehole image interpretation is considered in the stress polygon results if available.
+Overburden gradient [@traugott1997], pore-pressure [@Zhang20132; @Flemings2021], minimum horizontal stress [@Daines1982; @zoback1992], rock strength [@lal1999; @horsrud2001] and other parameters are calculated considering the given well-logging, deviation, and formation data. The maximum horizontal stress is estimated by applying stress-polygon [@ZOBACK2003] for every depth-sample. Borehole image interpretation is considered in the stress-polygon results if available.
 
-The calculation of tilted stress states using the given methodology requires the Euler angles $\alpha$, $\beta$ and $\gamma$. We calculate the Euler angles from geological data in terms of tilt azimuth and tilt angle of the stress tensor.
-
-The tilt of the stress tensor is calculated as follows:
[…]
