### 2025-02-27 841174b
Updated docs, adjust paper
