### 2025-03-27 de9ba4e
Update some plots and added biot to stress polygon calculations
