+++ paper/paper.md
@@ -1,10 +1,10 @@
 ---
-title: "Stresslog: A python package for modelling wellbore stability in inclined stress states"
+title: "Stresslog: A python package for modeling wellbore stability in inclined stress states"
 tags:
   - Python
   - Drilling
   - Geomechanics
-  - pore-pressure
+  - Pore-pressure
   - Mud loss
   - Wellbore stability
 authors:
@@ -23,11 +23,11 @@ bibliography: paper.bib
 
 # Summary
 
-This package is meant to be used by researchers and practitioners working in the field of geomechanics. It uses a collection of algorithms used to iteratively model the sate of stress underground, given a well log. The computations use 6 component stress tensor (as calculated using [@pevska1995]) which allows modelling of inclined wellbores in inclined states of stress. Stresslog estimates pore-pressure (with and without unloading effects) from sonic, resistivity and d-exponent, principal stresses, hoop stresses, compressive strength, tensile strength, the three moduli of elasticity and sanding tendency.
+This package is meant to be used by researchers and practitioners working in the field of geomechanics. It uses a collection of algorithms used to iteratively model the sate of stress underground, given a well log. The computations use 6 component stress tensor (as calculated using [@pevska1995]) which allows modeling of inclined wellbores in inclined states of stress. Stresslog estimates pore-pressure (with and without unloading effects) from sonic, resistivity and d-exponent, principal stresses, hoop stresses, compressive strength, tensile strength, the three moduli of elasticity and sanding tendency.
 
 # Statement of need
 
-Stresslog has been designed to help with pre-drill, post-drill and realtime geomechanical calculations. Stresslog is aimed at empowering researchers with a comprehensive 1D mechanical earth modelling tool that is freely available and which researchers can modify to apply their own methods when neccessary, while allowing practitioners to use industry-standard algorithms to calculate solutions and export well log data.
+Stresslog has been designed to help with pre-drill, post-drill and realtime geomechanical calculations. Stresslog is aimed at empowering researchers with a comprehensive 1D mechanical earth modeling tool that is freely available and which researchers can modify to apply their own methods when necessary, while allowing practitioners to use industry-standard algorithms to calculate solutions and export well log data.
 
 # Methodology
 
@@ -45,7 +45,7 @@ $$
 
 Where $R_s$ is the rotation matrix defined by Euler angles $\alpha$, $\beta$ and $\gamma$, in the NED reference frame.
 
-Considering the azimuth of maximum principal stress as $\alpha$, the above relations are used to optimise for the angles $\beta$ and $\gamma$. The technique proposed by [@pevska1995] starts with good estimates of the far field principal stresses, $\sigma_1$, $\sigma_2$ and $\sigma_3$, already rotated by the Euler Angles $\alpha$, $\beta$ and $\gamma$. Usually, however, what is available is an estimate of minimum horizontal stress, an estimate of the vertical stress, and an estimate of maximum horizontal stress. It is insufficient to simply rotate the tensor, as the rotated tensor will not have the correct vertical component. We optimise the principal stresses ($\sigma_1$, $\sigma_2$ and $\sigma_3$) such that the vertical and horizontal components of the tensor match the specified vertical and horizontal stresses.
[…]
