+++ paper/paper.bib
@@ -173,3 +173,10 @@
     url = {10.2118/56862-PA},
     eprint = {https://onepetro.org/DC/article-pdf/14/02/85/2590516/spe-56862-pa.pdf},
 }
+
+@misc{NorthernLightsJVDA,
+  title        = {Northern Lights JV DA},
+  howpublished = {\url{https://data.equinor.com/dataset/NorthernLights}},
+  note         = {Accessed July 2023},
+  year         = {2023}
+}
