+++ paper/paper.bib
@@ -174,7 +174,7 @@
     eprint = {https://onepetro.org/DC/article-pdf/14/02/85/2590516/spe-56862-pa.pdf},
 }
 
-@misc{NorthernLightsJVDA,
+@misc{northernlights,
   title        = {Northern Lights JV DA},
   howpublished = {\url{https://data.equinor.com/dataset/NorthernLights}},
   note         = {Accessed July 2023},
