+++ paper/paper.bib
@@ -176,7 +176,6 @@
 
 @misc{northernlights,
   title = {Northern Lights Dataset: Dataset from NO 31/5-7 (Eos) well},
-  author = {{Equinor ASA and the Northern Lights collaboration partners Total E\&P Norge AS and A/S Norske Shell}},
   year = {2020},
   url = {https://data.equinor.com},
   note = {Last update: 23 October 2020. Accessed July 2023. Available at https://data.equinor.com/dataset/NorthernLights after sign-in. Licensed under modified CC BY 4.0 (the licensed material may not be sold, and the license covers all data in the dataset).}
