### 2025-04-10 #7824 @mancellin
Sorry the scope of this paper is a bit too far from my current work for me to focus on it.
