### 2025-04-11 GeoArkadeep/Stresslog#1 @wkearn
Add supporting data for the Northern Lights example
Hi @GeoArkadeep, I'm working on my review over at https://github.com/openjournals/joss-reviews/issues/8036, and I noticed that it is somewhat challenging to replicate the Northern Lights example because one has to manually create the supporting CSV files (the ucs, formations and survey tables).

Would it make sense to include these CSV files in the repository or somewhere else where users could access them? It seems like Equinor [allows sharing](https://datavillagesa.blob.core.windows.net/disclaimers/Northern%20Lights%20data%20sharing.pdf) the data as long as every thing is properly attributed.
