### 2025-04-11 GeoArkadeep/Stresslog#1 @GeoArkadeep
Hi!
Sorry for the oversight on this. I have uploaded the supporting data here: https://github.com/GeoArkadeep/supporting-data-for-EOS-Northern-Lights

I hope the data is properly attributed, it would be of great help if you could offer feedback on the attribution portion.
