### 2025-04-19 8671ead
Update setup.py
Experimental support for python 3.7,3.8,3.9,3.12,3.13
