### 2025-04-19 de60aaf
Bumped version to 1.6.6
