+++ setup.py
@@ -3,7 +3,7 @@ import setuptools
 
 setuptools.setup(
     name="stresslog",
-    version="1.6.4",
+    version="1.6.6",
     description="Library for stress calculations from well logs",
     long_description=pathlib.Path("README.md").read_text(),
     long_description_content_type="text/markdown",
@@ -28,7 +28,7 @@ setuptools.setup(
         "Programming Language :: Python :: 3.13",
         "Topic :: Scientific/Engineering",
     ],
-    python_requires=">=3.10,<3.12",
+    python_requires=">=3.7,<=3.13",
     install_requires=["pandas>=2.0","numpy","scipy","welly","pint","matplotlib", "plotly","dlisio"],
     packages=setuptools.find_packages(),
     include_package_data=True,    
