### 2025-04-23 54ec1f3
Move setup.py into pyproject.toml
