+++ pyproject.toml
@@ -1,6 +1,48 @@
 [build-system]
-requires = [
-    "setuptools>=42",
-    "wheel",
-]
+requires = ["setuptools>=42", "wheel"]
 build-backend = "setuptools.build_meta"
+
+[project]
+name = "stresslog"
+version = "1.6.6"
+description = "Library for stress calculations from well logs"
+readme = "README.md"
+requires-python = ">=3.7,<=3.13"
+license = {text = "AGPL-3.0"}
+authors = [
+    {name = "Arkadeep Ghosh", email = "arkadeep_ghosh@rocklab.in"}
+]
+classifiers = [
+    "Development Status :: 6 - Mature",
+    "License :: OSI Approved :: GNU Affero General Public License v3",
+    "Intended Audience :: Science/Research",
+    "Programming Language :: Python :: 3.7",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Topic :: Scientific/Engineering"
+]
+dependencies = [
+    "pandas>=2.0",
+    "numpy",
+    "scipy",
+    "welly",
+    "pint",
+    "matplotlib",
+    "plotly",
+    "dlisio"
+]
+
+
+[project.urls]
+Homepage = "https://www.rocklab.in/stresslog"
+Documentation = "https://stresslog.readthedocs.io/en/stable/"
+Source = "https://github.com/GeoArkadeep/Stresslog"
+
+[tool.setuptools]
+packages = ["stresslog"]
+include-package-data = true
+
