### 2025-04-23 GeoArkadeep/Stresslog#3 @frank1010111
Documentation for testing
Hi! This is part of my [Joss review](https://github.com/openjournals/joss-reviews/issues/8036).

I have not been able to find the documentation for testing. It would be great to include how to run the automatic tests in your contributing guide.

It looks like you're using pytest with no extra plugins, but if I'm wrong or that changes, I'd like to see the optional dependencies. See, for instance, how welly's docs [cover development](https://github.com/agilescientific/welly/blob/main/docs/development.md) and how their pyproject has an [optional dependency section](https://github.com/agilescientific/welly/blob/main/pyproject.toml#L39) for testing.
