### 2025-04-23 GeoArkadeep/Stresslog#4 @frank1010111
Replicating EOS example
Hi! As part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/8036), I am trying to replicate the Equinor example.

I've run into the following issues

- [x] There's *a lot* of things printed out as the files are downloaded
- [x] pandas needs to be imported before it is used to load the surveys, formations,  ucs, and image log
- [x] `get_dlis_data` does not recognize that the DLIS file is in the subdirectory, apparently due to it stripping part of the path before loading the file
- [x] There are a lot of side effects, such as print statements and file creation in `lst.compute_geomech`. It would be better to have everything run silently and then allow the user to print calculations and create files as they feel necessary
- [x] There are warnings from `compute_geomech`, including welly, pyplot's renderer and pandas
- [x] It isn't immediately clear how we look at the borehole image after adding the directional survey
- [x] The plot in Iteration 4 is unlabeled, both axes and title
- [x] I don't know how the overlay created in Iteration 4
