### 2025-04-24 GeoArkadeep/Stresslog#4 @wkearn
All of what @frank1010111 noted has also come up for me when running the example.

In addition, the output I get does not seem to match the output in the documentation. For example, my output data frame from [Iteration 1](https://stresslog.readthedocs.io/en/latest/Examples.html#iteration-1-vertical-well) looks like

```
        DEPT  DTCO  DTSM  GR  NPHI  RLA1  RXO_HRLT  RHOZ  PEFLA      MD  ...  \
0        0.0   NaN   NaN NaN   NaN   NaN       NaN   NaN    NaN     0.0  ...
1        1.0   NaN   NaN NaN   NaN   NaN       NaN   NaN    NaN     1.0  ...
2        2.0   NaN   NaN NaN   NaN   NaN       NaN   NaN    NaN     2.0  ...
3        3.0   NaN   NaN NaN   NaN   NaN       NaN   NaN    NaN     3.0  ...
4        4.0   NaN   NaN NaN   NaN   NaN       NaN   NaN    NaN     4.0  ...
...      ...   ...   ...  ..   ...   ...       ...   ...    ...     ...  ...
2911  2911.0   NaN   NaN NaN   NaN   NaN       NaN   NaN    NaN  2911.0  ...
2912  2912.0   NaN   NaN NaN   NaN   NaN       NaN   NaN    NaN  2912.0  ...
2913  2913.0   NaN   NaN NaN   NaN   NaN       NaN   NaN    NaN  2913.0  ...
2914  2914.0   NaN   NaN NaN   NaN   NaN       NaN   NaN    NaN  2914.0  ...
2915  2915.0   NaN   NaN NaN   NaN   NaN       NaN   NaN    NaN  2915.0  ...
```

It seems like I end up calculating everything at a different set of depths? I have noticed that the output from

```python
string_las1 = lst.get_las_from_dlis('WL_RAW_AAC-ARLL-CAL-DEN-GR-NEU_RUN6_EWL_2.DLIS', aliases=alias, step=0.147)
vertwell = Well.from_las(string_las1)
vertwell.df()
```

is sampled at 1 m depth intervals unless I set the `step` argument very carefully.
