### 2025-04-24 GeoArkadeep/Stresslog#4 @GeoArkadeep
@wkearn , thank you for bringing this to my notice. This is a serious enough bug to warrant its own separate issue. Would you be kind enough to open up an issue focusing on this? Thank you very much.
