### 2025-04-24 GeoArkadeep/Stresslog#4 @GeoArkadeep
@frank1010111 , you have raised some excellent points. I am working on resolving these, hope to be able to fix them today.
