### 2025-04-24 4fa3710
Bugfixes and docfixes
Sorted the resampling issue (temporarily) by setting numpy version
Modified docs
