### 2025-04-24 ced701b
Update pyproject.toml
Forgot to update version
