+++ pyproject.toml
@@ -7,7 +7,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "stresslog"
-version = "1.6.6"
+version = "1.6.7"
 description = "Library for stress calculations from well logs"
 readme = { file = "README.md", content-type = "text/markdown" }
 requires-python = ">=3.9,<=3.13"
