### 2025-04-24 e7a514c
Update Examples.rst
