### 2025-04-25 GeoArkadeep/Stresslog#4 @GeoArkadeep
@frank1010111 I have updated the code and documentation with the changes desired. However not all points were fixed, I wish to address the remaining concerns here:
1) Side effects: It is true that a lot of files can be written using the method. This includes various configuration files, plots, as well as writing out the calculated dataframe to csv and las files. This is very easily preventable by setting writeFile (and debug) to False, at which point only configuration files will be written (and terminal output skipped). The configuration files do not **_need_** to be written or present on disk for the program to execute correctly, but missing the configuration files will cause the program to revert to default alias and plot settings. The point of the configuration files is to let users new to python (but perhaps not to geomechanics) an easy way to update the log aliases and plot config without having to modify the code. It is a feature, not a bug, and will be added to the documentation soon.

2) Overlay code: The overlaying of the synthetic borehole image onto the image log, is **not a feature offered by Stresslog.** It is done manually. However, for sake of the paper, the code actually implementing this has always been present at [docs/figures/combiner.py
](https://github.com/GeoArkadeep/Stresslog/blob/master/docs/Figures/combiner.py)
I hope this sufficiently addresses all the raised issues.
