### 2025-04-26 903d049
Workaround to numpy>1.6.4 issue
Added method get_well_from_dlis which resolves the resampling issue by explicitly using a basis.
