### 2025-04-26 c6cfffa
Updated docstrings, bumped version
