### 2025-04-26 eb9e664
Update syntheticLogs.py
Fixed the hardcoded file path left in by mistake
