### 2025-04-26 f529e00
Update geomechanics.py
added warning suppression
