### 2025-04-26 9d579eb
Update geomechanics.py
Suppressed ALL warinings. Not what I wanted, but it is what is easy, and also what was asked for, so here we go.
