### 2025-04-30 c849f74
Bump version
Final changes for the three issues
