### 2025-04-30 GeoArkadeep/Stresslog#4 @GeoArkadeep
I am adding a github comparison tracking all the changes to all the issues listed here, as well as #5 and #3. Hope this will make it easier to track changes.

https://github.com/GeoArkadeep/Stresslog/compare/de60aaf908a3cf8a5273b2eb3ae510dc169a7939...cba3c167c7ec805c4b95a9a6c907f796decf0f96
