### 2025-04-30 cba3c16
FIxed nugs and bump version
Added docstring for get_well_from_dlis
