### 2025-04-30 GeoArkadeep/Stresslog#3 @GeoArkadeep
@frank1010111 here is a comparison showing the changes to the testing documentation: https://github.com/GeoArkadeep/Stresslog/compare/de60aaf908a3cf8a5273b2eb3ae510dc169a7939...c849f74b5dce68065ac6c4fae8122dc45a63d19f

The relevant changes are only in the Introduction.rst file, please ignore the other changes for this issue.
