### 2025-05-01 GeoArkadeep/Stresslog#3 @GeoArkadeep
For sure. I'll create dedicated issue branches for future issues and use pull requests to merge the changes. As you can tell I am new to repo management.

I'm gonna go ahead and close this issue now, as I believe the subject has been resolved.
