+++ docs/source/Examples.rst
@@ -200,25 +200,21 @@ Our first analysis assumes a vertical well:
    print(output[0])
    print(list(output[0]))
    """
-                     DEPT           DTCO  ...  Shear_Modulus     Bulk_Modulus
-   0         0.0000000000            NaN  ...   0.0000000000     0.0000000000
-   1         0.1470000000            NaN  ...   0.0000000000     0.0000000000
-   2         0.2940000000            NaN  ...   0.0000000000     0.0000000000
-   3         0.4410000000            NaN  ...   0.0000000000     0.0000000000
-   4         0.5880000000            NaN  ...   0.0000000000     0.0000000000
-   ...                ...            ...  ...            ...              ...
-   18922  2781.5339999988  87.2171899945  ...   1.8637329689  5133.1104753863
-   18923  2781.6809999988  87.7187699958  ...   1.8287958424  4848.8212822883
-   18924  2781.8279999988  88.2238099957  ...   1.7944750138  4796.3186612756
-   18925  2781.9749999988  88.4480199981  ...   1.7795074291  4957.0108813337
-   18926  2782.1219999988  88.3849000005  ...   0.0000000000     0.0000000000
+                  DEPT  DTCO  ...  Shear_Modulus     Bulk_Modulus
+   0         0.0000000000   NaN  ...   0.0000000000     0.0000000000
+   1         0.1470000000   NaN  ...   0.0000000000     0.0000000000
+   2         0.2940000000   NaN  ...   0.0000000000     0.0000000000
+   3         0.4410000000   NaN  ...   0.0000000000     0.0000000000
+   4         0.5880000000   NaN  ...   0.0000000000     0.0000000000
+   ...                ...   ...  ...            ...              ...
+   19828  2914.7159999987   NaN  ...   2.2818260018  6776.4871193190
+   19829  2914.8629999987   NaN  ...   2.2820489493  6777.2354994299
+   19830  2915.0099999987   NaN  ...   2.2822719005  6777.9839018296
+   19831  2915.1569999987   NaN  ...   2.2824948556  6778.7323265105
+   19832  2915.3039999987   NaN  ...   0.0000000000     0.0000000000
 
-   [18927 rows x 38 columns]
-   ['DEPT', 'DTCO', 'DTSM', 'GR', 'NPHI', 'RLA1', 'RXO_HRLT', 'RHOZ', 'PEFLA',
-   'MD', 'TVDM', 'INCL', 'AZIM', 'ShaleFlag', 'RHO', 'OBG_AMOCO', 'DTCT', 'PP_GRADIENT',
-   'SHmin_DAINES', 'SHmin_ZOBACK', 'FracGrad', 'FracPressure', 'GEOPRESSURE', 'SHmin_PRESSURE', 'SHmax_PRESSURE',
-   'MUD_PRESSURE', 'OVERBURDEN_PRESSURE', 'HYDROSTATIC_PRESSURE', 'MUD_GRADIENT', 'S0_Lal', 'S0_Lal_Phi', 'UCS_Horsud',
-   'UCS_Lal', 'Poisson_Ratio', 'ML90', 'Youngs_Modulus', 'Shear_Modulus', 'Bulk_Modulus']
+   [19833 rows x 40 columns]
[…]
