### 2025-05-16 GeoArkadeep/Stresslog#7 @wkearn
Discuss the state of the field in the JOSS paper
One thing that is missing from the JOSS manuscript is a discussion of how Stresslog compares to existing software. In particular, I would like to know more about how the design of Stresslog reflects the needs of its user community, how it responds to gaps in or challenges posed by existing software, and how it operates with other tools and sources of information. The current paper focuses a lot on the modeling techniques used by the software, and these are details that I'm not sure are necessary. It would be nice to see in the paper a strong argument for why a user should develop a workflow using Stresslog.
