### 2025-05-19 1a5bc02
Resolve side effects
File writing side effects have been resolved.
