### 2025-05-19 GeoArkadeep/Stresslog#4 @GeoArkadeep
The file writing side effects have now been made optional.

The commit implementing this: 1a5bc022f0b3d32decdbf06eaa42fc35b43aa21e

The documentation has also been updated to reflect this.
