+++ paper/paper.bib
@@ -180,3 +180,5 @@
   url = {https://data.equinor.com},
   note = {Last update: 23 October 2020. Accessed July 2023. Available at https://data.equinor.com/dataset/NorthernLights after sign-in. Licensed under modified CC BY 4.0 (the licensed material may not be sold, and the license covers all data in the dataset).}
 }
+
+@article{Yu2018, doi = {10.21105/joss.00992}, url = {https://doi.org/10.21105/joss.00992}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {30}, pages = {992}, author = {Hao Yu}, title = {PyGeoPressure: Geopressure Prediction in Python}, journal = {Journal of Open Source Software} }
