### 2025-05-19 GeoArkadeep/Stresslog#7 @GeoArkadeep
Thank you for pointing out this omission, I have now added these in the paper.

https://github.com/GeoArkadeep/Stresslog/compare/34c0b70f1032770d95344dedb135148c7c2f4405...8bb058e0dd964f0ee01f550444808d2630a18139

I want to take a moment and defend the inclusion of the modeling techniques in detail. See the derived equation is (in my judgement) part of my "scholarly effort" required for JOSS papers. In particular, because it makes it more "likely to be cited by other researchers" (in my personal opinion). Further, the equation was derived using https://github.com/GeoArkadeep/Stresslog/blob/master/paper/fracgrad_equation.py, so it is part of the repository (though the derivation is not part of the package, the formula definitely is). Without this, it would be just another implementation of existing formulae.

Thus, if not a huge problem, I would like to keep the methodology section intact in the paper.
