### 2025-05-20 5d2a77d
Update paper.md
Fix typo
