### 2025-05-20 a0228a9
Update fracgrad_equation.py
