### 2025-05-27 GeoArkadeep/Stresslog#4 @frank1010111
I've run through the EOS example again and it is much improved. It is still a little confusing to see the figures accompanying the geomechanical computations without code to show them.
