### 2025-05-28 GeoArkadeep/Stresslog#4 @GeoArkadeep
I'm not sure I completely follow. Is this about the overlay image? Or all the images returned by the function?
Any suggestions about how they should behave?
