### 2025-05-28 GeoArkadeep/Stresslog#4 @GeoArkadeep
Are we talking about something like plt.show()?
