### 2025-06-02 GeoArkadeep/Stresslog#7 @wkearn
I understand more now about how that methodology section fits into things. It's still a little stylistically weird to me to have a half-page equation in a 3 page paper, but if you think it is important, it is all right to keep it.

The state of the field does now give an argument for what Stresslog provides beyond existing open-source and commercial software packages. This section could always be more comprehensive. It might be a good idea, for instance, to provide citations for the `dlisio`, `lasio` and `welly` packages, which you depend on directly or indirectly, and which are clearly related to your domain, even if they are not geomechanical modelling software.
