### 2025-06-02 #8036 @wkearn
@willgearty: Sorry for the somewhat slow progress here. I think the paper has improved with the changes discussed in GeoArkadeep/Stresslog#7, and I just responded there with a few brief suggestions that I think will finish up the paper section for me.

There is still some room for improvement in the testing situation, and I am curious what @frank1010111 thinks on this front, having noted it [earlier](https://github.com/openjournals/joss-reviews/issues/8036#issuecomment-2861391514). There are a few test cases in the repository, and the comprehensive example from the documentation now reproduces correctly and can be a good manual test case, so I think we currently fall under the "OK" category [here](https://joss.readthedocs.io/en/latest/review_criteria.html#tests).
