### 2025-06-03 cac81be
test initial idea
