### 2025-06-03 8e073ea
Added the other plots and checked function
Now shows Stress Polygon, Directional Stability Plot, Synthetic Borehole Image, Sanding Risk Plot and Well Plot if display=True
