### 2025-06-03 GeoArkadeep/Stresslog#4 @GeoArkadeep
I have added (optional) plt.show() to key figures during the process: https://github.com/GeoArkadeep/Stresslog/pull/8

Please let me know if this is good enough.

Docs have been updated https://github.com/GeoArkadeep/Stresslog/commit/7e716486fd125ee741efc5d048f17a9b3faefc50
