### 2025-06-03 GeoArkadeep/Stresslog#7 @GeoArkadeep
I would to include citations for dlisio, lasio, and welly. However I am unable to find either a bibtex citation or an archive with doi for these libraries.

What I can find citations for are scipy, numpy and matplotlib

I would love to include these as well as lasio, dlisio and welly. Would love to know your thoughts/advice on this.
