### 2025-06-03 5c39f0b
Updated scipy citation
