### 2025-06-03 GeoArkadeep/Stresslog#7 @GeoArkadeep
Meanwhile, I have gone ahead and cited scipy, and mentioned dlisio, lasio and welly in the paper (though I could not properly cite them, as explained). https://github.com/GeoArkadeep/Stresslog/compare/7e716486fd125ee741efc5d048f17a9b3faefc50...83c3c9877f010590c2e0f81329e1a5071d25da7b

Please let me know if this is sufficient.

In particular I had to put it in the Summary section, because I could not think of where else to put it (it does not quite fit there as well, but had to put it somewhere. These have always been a part of the readme, so if you think it does not fit well in the summary section, perhaps the readme alone would suffice? Please let me know).
