+++ tests/test_package.py
@@ -7,7 +7,7 @@ See the GNU Affero General Public License for more details: <https://www.gnu.org
 import pytest
 import stresslog as lst
 import pandas as pd
-
+import numpy as np
 
 def test_integration_basic():
 # Test case 1: Basic well log generation and plotting
@@ -29,12 +29,17 @@ def test_integration_doi():
     assert output is not None, "plotPPzhang with attributes and DOI did not return an output"
 
 def test_integration_advanced():
-    # Test case 3: Shallow angle well and custom TECB
+    # Test case 3: Shallow angle well and custom TECB. now with Formations, Imagelog and UCS data
+
+    ucs = pd.read_csv('https://raw.githubusercontent.com/GeoArkadeep/supporting-data-for-EOS-Northern-Lights/main/UCSdata.csv')
+    formations = pd.read_csv('https://raw.githubusercontent.com/GeoArkadeep/supporting-data-for-EOS-Northern-Lights/main/NorthernLights-31_5-7.csv')
+    imagelog = pd.read_csv('https://raw.githubusercontent.com/GeoArkadeep/supporting-data-for-EOS-Northern-Lights/main/31_5-7_Image.csv')
+    
     test_m_depth=2000
     attrib = [10, -200, 0, 0, 0, 0, 0, 0]
     well = lst.create_random_well(kb=35, gl=-200, step=4, seed=33)
     dev = lst.getwelldev(wella=well, kickoffpoint=300, final_angle=10, rateofbuild=0.2, azimuth=270)
-    output = lst.compute_geomech(dev, attrib=attrib, doi=2625, tecb=0.55, writeFile=False)
+    output = lst.compute_geomech(dev, attrib=attrib, doi=2625, tecb=0.55, display=True, forms = formations, UCSs = ucs, flags = imagelog, debug = True)
     
     index = lst.find_nearest_depth(output[0]['MD'].to_numpy(),test_m_depth)[0]
     
@@ -44,5 +49,31 @@ def test_integration_advanced():
     assert abs(output[0]['FracPressure'].to_numpy()[index] - 9402.524110909559) < 0.001, f"Frac Pressure check fail: {output[0]['FracPressure'].to_numpy()[index]}"
     print(f"Frac Pressure check passed: {output[0]['FracPressure'].to_numpy()[index]}psi")
 
-#if __name__=="__main__":
-#    integrationTest()
+def test_analog():
+    #Test case 4: Test analog well
+    formations = pd.read_csv('https://raw.githubusercontent.com/GeoArkadeep/supporting-data-for-EOS-Northern-Lights/main/NorthernLights-31_5-7.csv')
+    well = lst.create_random_well(kb=35, gl=-200, step=4, seed=33)
+    dev = lst.getwelldev(wella=well, kickoffpoint=300, final_angle=10, rateofbuild=0.2, azimuth=270)
+    
+    new_forms = formations.rename(columns={"Top TVD": "TopTVD", "Formation Name": "Name"})[["TopTVD", "Name"]]
+    target_forms = formations.rename(columns={"Top TVD": "TopTVD", "Formation Name": "Name"}).assign(TopTVD=lambda df: df["TopTVD"] + 150)[["TopTVD", "Name"]]
+    analogwell,analoglas, analogdev = lst.get_analog(dev,new_forms, target_forms, kb=40,gl=-200,dev=None,kop=0,ma=0,rob=0.1,azi=0,debug=False,td=5000)
+    
+    df_analog = analogwell.df()
+    dev_df = dev.df() # 'dev' is the source well passed to lst.get_analog
+
[…]
