### 2025-06-04 GeoArkadeep/Stresslog#7 @wkearn
I think you can use an `@misc` entry in your bibtex file with some information about the software and authors even if they don't have a formal publication of the software. [welly](https://github.com/agilescientific/welly/blob/b44e90b743d954802f36a98f0fd35095e110d22d/CITATION.cff) and [lasio](https://github.com/kinverarity1/lasio/blob/9e32ea67f8589955e62d979251c514d290e00b9c/CITATION.cff) at least have CITATION.cff files that suggest how they would like to be cited, and you could probably do something similar for dlisio.

Putting this information in the Summary section works. Alternatively, you could have a short paragraph in Methodology that describes the input data required and references the formats that you can read data from. Either way, I think it is useful to have that information in the paper as a way of acknowledging the work of those developers.
