### 2025-06-04 0e31af3
Updated welly lasio and dliso citations
