+++ paper/paper.bib
@@ -205,3 +205,27 @@
   pages   = {261--272},
   doi     = {10.1038/s41592-019-0686-2},
 }
+
+@software{welly,
+  author       = {Welly contributors},
+  title        = {Welly, a Python package},
+  year         = {2021},
+  url = {https://github.com/agilescientific/welly},
+}
+
+@software{dlisio,
+  author       = {Equinor ASA and contributors},
+  title        = {Dlisio: Python library for working with the well log formats Digital Log Interchange Standard (DLIS V1) and Log Information Standard (LIS79)},
+  year         = {2025},
+  url = {https://github.com/equinor/dlisio},
+}
+
+@software{Inverarity_Lasio_2023,
+author = {Inverarity, Kent and {Lasio contributors}},
+license = {MIT},
+month = may,
+title = {{Lasio}},
+url = {https://github.com/kinverarity1/lasio},
+version = {0.31},
+year = {2023}
+}
