+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 # Summary
 
-This package is meant to be used by researchers and practitioners working in the field of geomechanics. It uses a collection of algorithms used to iteratively model the sate of stress underground, given a well log. The computations use 6 component stress tensor, as calculated using [@pevska1995], which allows modeling of inclined wellbores in inclined states of stress. Under the hood, Stresslog uses SciPy [@SciPy] for optimizations, dlisio and welly to read .dlis files and .las files (and other well data), respectively, and lasio to write .las files.
+This package is meant to be used by researchers and practitioners working in the field of geomechanics. It uses a collection of algorithms used to iteratively model the sate of stress underground, given a well log. The computations use 6 component stress tensor, as calculated using [@pevska1995], which allows modeling of inclined wellbores in inclined states of stress. Under the hood, Stresslog uses SciPy [@SciPy] for optimizations, dlisio [@dlisio] and welly [@welly] to read .dlis files and .las files (and other well data), respectively, and lasio[@Inverarity_Lasio_2023] to write .las files.
 
 # Statement of need
 
