+++ paper/paper.bib
@@ -207,14 +207,14 @@
 }
 
 @software{welly,
-  author       = {Welly contributors},
+  author       = {{Welly contributors}},
   title        = {Welly, a Python package},
   year         = {2021},
   url = {https://github.com/agilescientific/welly},
 }
 
 @software{dlisio,
-  author       = {Equinor ASA and contributors},
+  author       = {{Equinor ASA \& dlisio contributors}},
   title        = {Dlisio: Python library for working with the well log formats Digital Log Interchange Standard (DLIS V1) and Log Information Standard (LIS79)},
   year         = {2025},
   url = {https://github.com/equinor/dlisio},
