### 2025-06-04 GeoArkadeep/Stresslog#7 @GeoArkadeep
Thank you for this great advice! I am sorry for overlooking the .cff files. I have updated the citations for welly, lasio and dlisio.

I have not added anything to the methodology section as we are already over the 1000 word limit for JOSS papers, I am trying to keep it as concise as possible.

I hope this is acceptable.
