### 2025-06-05 #8036 @frank1010111
Well, yes, that is the substance of my suggestion. Unit tests should test that valid outputs come from valid inputs and invalid inputs result in errors. Regression tests should ensure that code changes don't lead to known-to-be-good outputs turning into bad outputs. Integration tests should show that functions work well together, and these should all be automatically run. Code coverage is just one metric for judging code completeness.
