### 2025-06-06 GeoArkadeep/Stresslog#4 @frank1010111
I'm happy with where they are after your changes.
