### 2025-06-06 GeoArkadeep/Stresslog#4 @GeoArkadeep
I hope this resolves all the sub-problems encountered here.

I am therefore closing the issue, but if there are other concerns, please feel free to re-open the issue (or create new ones).

Solving the problems shared here led to improvement of the quality of the package and I am thus very thankful that these were brought up.
