### 2025-07-08 GeoArkadeep/Stresslog#7 @frank1010111
I think it would be productive to review how the paper is structured. They are meant to be quite different from traditional papers both in length and required sections: https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain

There isn't actually any requirement for a methodology section, and it is probably better put in the documentation rather than the paper itself. That would free up space to compare to other works, talk about the intended research purpose, etc.
