### 2025-07-08 #8036 @frank1010111
Excellent question. I was curious how the paper would change after @wkearn 's suggestions, so that would be a good starting point.

Then,  I don't quite know where to offer specific suggestions on the tests. They aren't really unit tests, per se, but I'm unclear on the JOSS expectations regarding testing. I wish I could articulate it better.
