+++ tests/test_units_stresslog.py
@@ -1,0 +1,451 @@
+import pytest
+import numpy as np
+import stresslog
+
+# --- Test Configuration ---
+RTOL = 0.01  # Relative Tolerance
+ATOL = 0.051  # Absolute Tolerance for 1-decimal rounding
+# ------------------------
+# --- Phase 1: Scalar Function Regression Tests ---
+
+@pytest.mark.parametrize("inputs, expected_output", [
+    ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
+ 6000.0]), 'porepressures': np.array([909.2, 3801.8, 6297.8, 7349.6, 9262.3, 12810.5, 13001.5, 14592.5, 18153.9,
+ 20549.0])}, 2.1),
+    ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
+ 6000.0]), 'porepressures': np.array([2366.9, 4223.2, 7577.7, 6020.6, 9510.0, 13165.1, 12597.8, 15780.9,
+ 17919.6, 19066.0])}, 2.2),
+    ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
+ 6000.0]), 'porepressures': np.array([1097.2, 5174.0, 5824.9, 7535.2, 8096.7, 11049.8, 13712.1, 16220.4,
+ 16702.5, 20602.8])}, 2.1),
+    ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
+ 6000.0]), 'porepressures': np.array([1703.4, 2977.0, 4788.5, 8110.4, 9125.1, 11923.2, 13231.0, 16285.9,
+ 19507.4, 19945.5])}, 2.2),
+    ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
+ 6000.0]), 'porepressures': np.array([1872.8, 3607.6, 6056.2, 7459.8, 10371.3, 11828.4, 14590.9, 14893.1,
+ 18591.2, 21427.9])}, 2.2),
+])
+def test_compute_optimal_gradient(inputs, expected_output):
+    """Regression test for stresslog.compute_optimal_gradient."""
+    actual_output = stresslog.compute_optimal_gradient(**inputs)
+    assert actual_output == pytest.approx(expected_output, rel=RTOL, abs=ATOL, nan_ok=True)
+
+
+@pytest.mark.parametrize("inputs, expected_output", [
+    ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
+ 6000.0]), 'porepressures': np.array([1507.1, 4182.9, 4142.0, 8134.4, 9551.1, 11783.9, 12754.1, 16508.1,
+ 19277.5, 19590.3]), 'gradient': 1.0}, 4045.6),
+    ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
+ 6000.0]), 'porepressures': np.array([2962.0, 3311.7, 4687.8, 7435.9, 9963.7, 12115.1, 14161.2, 14856.5,
+ 18913.4, 19173.5]), 'gradient': 2.3}, -214.7),
[…]
