### 2025-07-24 8cf6ce5
Update test_units_stresslog.py
