+++ tests/test_units_stresslog.py
@@ -10,20 +10,20 @@ ATOL = 0.051  # Absolute Tolerance for 1-decimal rounding
 
 @pytest.mark.parametrize("inputs, expected_output", [
     ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
- 6000.0]), 'porepressures': np.array([909.2, 3801.8, 6297.8, 7349.6, 9262.3, 12810.5, 13001.5, 14592.5, 18153.9,
- 20549.0])}, 2.1),
+ 6000.0]), 'porepressures': np.array([3200.0, 3464.8, 6070.9, 8790.7, 10603.2, 12037.5, 13928.6, 15543.7,
+ 18926.5, 21039.8])}, 2.3),
     ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
- 6000.0]), 'porepressures': np.array([2366.9, 4223.2, 7577.7, 6020.6, 9510.0, 13165.1, 12597.8, 15780.9,
- 17919.6, 19066.0])}, 2.2),
+ 6000.0]), 'porepressures': np.array([1108.1, 2316.3, 6557.0, 6792.8, 11063.0, 12879.0, 15308.5, 17338.0,
+ 16869.5, 20910.7])}, 2.2),
     ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
- 6000.0]), 'porepressures': np.array([1097.2, 5174.0, 5824.9, 7535.2, 8096.7, 11049.8, 13712.1, 16220.4,
- 16702.5, 20602.8])}, 2.1),
+ 6000.0]), 'porepressures': np.array([2320.9, 2439.8, 5828.6, 8851.9, 9860.2, 11649.9, 11761.3, 16160.0,
+ 18580.8, 20585.7])}, 2.2),
     ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
- 6000.0]), 'porepressures': np.array([1703.4, 2977.0, 4788.5, 8110.4, 9125.1, 11923.2, 13231.0, 16285.9,
- 19507.4, 19945.5])}, 2.2),
+ 6000.0]), 'porepressures': np.array([612.8, 3622.1, 5396.3, 6268.1, 9224.5, 12296.7, 14860.8, 15723.8, 18909.8,
+ 20765.6])}, 2.2),
     ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
- 6000.0]), 'porepressures': np.array([1872.8, 3607.6, 6056.2, 7459.8, 10371.3, 11828.4, 14590.9, 14893.1,
- 18591.2, 21427.9])}, 2.2),
+ 6000.0]), 'porepressures': np.array([2700.1, 3690.9, 6820.5, 8388.6, 10468.3, 10381.0, 12522.3, 14158.6,
+ 17927.8, 19984.7])}, 2.1),
 ])
 def test_compute_optimal_gradient(inputs, expected_output):
     """Regression test for stresslog.compute_optimal_gradient."""
@@ -33,20 +33,20 @@ def test_compute_optimal_gradient(inputs, expected_output):
 
[…]
