### 2025-07-24 8b4d47e
Create ci.yml
Initial commit testing CI workflow
