### 2025-07-24 #8036 @GeoArkadeep
Sorry for the delay, I was completely swamped with this other project I am working on.

I have added [unit tests](https://github.com/GeoArkadeep/Stresslog/blob/master/tests/test_units_stresslog.py).

I have also added a workflow such that both the integration and unit tests run anytime a pull request is made.
