### 2025-07-24 GeoArkadeep/Stresslog#7 @GeoArkadeep
One of the key problems of removing the methodology section is that most of the key geo-science citations will have to be skipped. In addition to the concerns I have raised above.

There is nothing in the [guidelines](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) that preclude the methodology section, as long as the overall paper length is within limits.

I have already included comparison to other works and research applications. Like I have nothing more to add to these sections even if length of the paper was not limited.

The usage instructions are present in the documentation, and the science portion is in the paper. I would like to keep it that way. This allows prospective users/researchers to judge whether this work meets their needs just by reading the paper, if they find it useful, they can use the documentation to see how to use it.
