### 2025-08-13 #8036 @willgearty
Sorry for the delay, I was travelling the last few weeks. Thanks for implementing the unit tests and automated workflow @GeoArkadeep! I think that resolves some of the reviewers concerns.

It looks like the remaining issues are still being discussed in https://github.com/GeoArkadeep/Stresslog/issues/7.
