### 2025-08-13 #8036 @GeoArkadeep
That's right. It would be very helpful if you (@willgearty) weighed in on the situation please. I think it would be very helpful if the JOSS position on this was made clear.
