### 2025-08-13 GeoArkadeep/Stresslog#7 @willgearty
For a larger project, the details in the current Methodology section might be published as a whole separate scientific article, but given how short it is (and since there is no plan to publish any separate scientific articles), I think it is fine to include here, and I'd agree with @GeoArkadeep that it makes more sense to include it in the paper rather than in the documentation (although that's just an opinion). Here is an example of a recently published JOSS paper which includes a similar section of about a similar length: https://joss.theoj.org/papers/10.21105/joss.07320.

That being said, I do agree with @wkearn and @frank1010111 that the Statement of Need section could be expanded a bit, in particular with more details about "[placing] it in the context of related work". See the above published paper for an example of a much larger Statement of Need section. I don't this paper requires a Statement of Need that is quite as long, but hopefully it exemplifies the kinds of details that should be in this section. As @wkearn suggested above, it may make sense to move some of the now-cited python packages into this section. It would also be good to cite any other python packages that do anything similar to this one (to provide evidence for why this python package is needed) as suggested by @frank1010111.
