### 2025-08-13 GeoArkadeep/Stresslog#7 @GeoArkadeep
I will go through this paper in detail and apply the required changes.
