### 2025-08-22 71c9ce9
Updateed paper
