+++ paper/paper.bib
@@ -229,3 +229,71 @@ url = {https://github.com/kinverarity1/lasio},
 version = {0.31},
 year = {2023}
 }
+
+@software{plotly_py,
+  author       = {Kruchten, Nicolas and Seier, Andrew and Parmer, Chris},
+  title        = {An interactive, open-source, and browser-based graphing library for Python},
+  version      = {5.24.1},
+  date         = {2024-09-12}s,
+  doi          = {10.5281/zenodo.14503524},
+  url          = {https://github.com/plotly/plotly.py},
+}
+
+@software{pandas_software,
+  author       = {{The pandas development team}},
+  title        = {pandas-dev/pandas: Pandas},
+  abstract     = {Pandas is a powerful data structures for data analysis, time series, and statistics.},
+  doi          = {10.5281/zenodo.3509134},
+  url          = {https://pandas.pydata.org/},
+  version      = {},
+  license      = {BSD-3-Clause},
+  note         = {Available at \url{https://github.com/pandas-dev/pandas}},
+}
+
+@article{2020NumPy-Array,
+  author  = {Harris, Charles R. and Millman, K. Jarrod and
+            van der Walt, Stéfan J and Gommers, Ralf and
+            Virtanen, Pauli and Cournapeau, David and
+            Wieser, Eric and Taylor, Julian and Berg, Sebastian and
+            Smith, Nathaniel J. and Kern, Robert and Picus, Matti and
+            Hoyer, Stephan and van Kerkwijk, Marten H. and
+            Brett, Matthew and Haldane, Allan and
+            Fernández del Río, Jaime and Wiebe, Mark and
+            Peterson, Pearu and Gérard-Marchant, Pierre and
+            Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and
+            Abbasi, Hameer and Gohlke, Christoph and
+            Oliphant, Travis E.},
+  title   = {Array programming with {NumPy}},
+  journal = {Nature},
+  year    = {2020},
+  volume  = {585},
+  pages   = {357–362},
+  doi     = {10.1038/s41586-020-2649-2}
+}
+
+@inproceedings{mckinney2010data,
+  author       = {McKinney, Wes},
+  title        = {Data Structures for Statistical Computing in Python},
+  booktitle    = {Proceedings of the 9th Python in Science Conference},
+  editor       = {van der Walt, Stéfan and Millman, Jarrod},
+  year         = {2010},
+  pages        = {56--61},
+  doi          = {10.25080/Majora-92bf1922-00a},
+  url          = {https://conference.scipy.org/proceedings/scipy2010/mckinney.html},
+  address      = {Austin, TX, USA},
+}
+
+@Article{Hunter:2007,
+  Author    = {Hunter, J. D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing in Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  application development, interactive scripting, and publication-quality
+  image generation across user interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = 2007
+}
