### 2025-08-22 e5bf2bf
Update paper.bib
