+++ paper/paper.bib
@@ -234,7 +234,7 @@ year = {2023}
   author       = {Kruchten, Nicolas and Seier, Andrew and Parmer, Chris},
   title        = {An interactive, open-source, and browser-based graphing library for Python},
   version      = {5.24.1},
-  date         = {2024-09-12}s,
+  date         = {2024-09-12},
   doi          = {10.5281/zenodo.14503524},
   url          = {https://github.com/plotly/plotly.py},
 }
