### 2025-08-22 237a015
Update paper.md
