### 2025-08-22 4b3547b
Update paper.bib
