+++ paper/paper.bib
@@ -239,15 +239,15 @@ year = {2023}
   url          = {https://github.com/plotly/plotly.py},
 }
 
-@software{pandas_software,
-  author       = {{The pandas development team}},
+@software{the_pandas_development_team_2025_16918803,
+  author       = {The pandas development team},
   title        = {pandas-dev/pandas: Pandas},
-  abstract     = {Pandas is a powerful data structures for data analysis, time series, and statistics.},
-  doi          = {10.5281/zenodo.3509134},
-  url          = {https://pandas.pydata.org/},
-  version      = {},
-  license      = {BSD-3-Clause},
-  note         = {Available at \url{https://github.com/pandas-dev/pandas}},
+  month        = aug,
+  year         = 2025,
+  publisher    = {Zenodo},
+  version      = {v2.3.2},
+  doi          = {10.5281/zenodo.16918803},
+  url          = {https://doi.org/10.5281/zenodo.16918803},
 }
 
 @article{2020NumPy-Array,
