### 2025-08-22 53d0c4a
Update paper.bib
