+++ paper/paper.bib
@@ -240,7 +240,7 @@ year = {2023}
 }
 
 @software{the_pandas_development_team_2025_16918803,
-  author       = {The pandas development team},
+  author       = {{The pandas development team}},
   title        = {pandas-dev/pandas: Pandas},
   month        = aug,
   year         = 2025,
