### 2025-08-24 GeoArkadeep/Stresslog#7 @GeoArkadeep
I have expanded the Statement of need section, expanded the citations, and added a brief note about the type of data required in the Methodology section.
