### 2025-08-26 GeoArkadeep/Stresslog#7 @willgearty
The statement of need section looks to be much improved! What do you think @wkearn @frank1010111?
