### 2025-08-27 GeoArkadeep/Stresslog#7 @frank1010111
I'm now happy with the paper. Thank you, GeoArkadeep.
