### 2025-08-27 GeoArkadeep/Stresslog#9 @frank1010111
Fix a few issues in the API level docs
I have a few documentation suggestions as part of this JOSS review https://github.com/openjournals/joss-reviews/issues/8036.

When reviewing the API docs, I found a few formatting issues:

- [x] `add_curves` is not properly parsed by sphinx
- [x] The return statement for `compute_geomech` is, I think, meant to be an ordered list
- [x] `datasets_to_las` just provides an Agile Scientific copyright notice
- [x] `draw`'s notes section is, I think, meant to return an unordered list for the last paragraph
- [x] `drawSP`'s notes section also looks meant to show an unordered list
- [x] `find_TVD` has no docstring
- [x] `getColumnHeights` looks like the return section ought to have an unordered list
- [x] `getHoop`'s notes section is meant to have an unordered list
- [x] `getSP`'s `flag` parameter is meant to have a list and its return statement as well
- [x] `getVertical`'s return section is meant to have a list
- [x] `get_OBG_pascals_vec`'s parameters are not parsed correctly
- [x] `get_Shmin_grad_Daine_ppg`'s parameters are not parsed correctly
- [x] `get_Shmin_grad_Daine_ppg_vec`'s parameters are not parsed correctly
- [x] `get_analog` looks like it is meant to include an ordered list in the description
- [x] `get_principal_stress` seems to be incorrectly parsed
- [x] `plot_logs_labels` has the parameters and return sections incorrectly parsed
- [x] `plotfrac`'s return section is missing a bullet point
- [x] `remove_curves`'s parameters are not parsed
