### 2025-08-27 GeoArkadeep/Stresslog#10 @frank1010111
Mutable objects in function signatures
As part of my code review for [JOSS](https://github.com/openjournals/joss-reviews/issues/8036), I found a few functions use mutable objects as default values.

The examples I found are lists and dictionaries. This can lead to [weird issues](https://www.reddit.com/r/Python/comments/opb7hm/do_not_use_mutable_objects_as_default_arguments/).

- [x] `compute_geomech` uses lists for mwvalues, flowgradvals, fracgradvals, flowpsivals, fracpsivals, attrib, and program_option
- [x] `compute_geomech` also uses a dictionary for unitdict
- [x] `create_random_well` uses a list for drop
- [x] `plot_logs_labels` uses a dictionary for details
- [x] `datasets_to_las` uses a dictionary for custom_units
