### 2025-08-27 ad3190a
Mutable defaultsfootgun safing
