### 2025-08-27 8ae8a36
Update geomechanics.py
