### 2025-08-28 GeoArkadeep/Stresslog#10 @GeoArkadeep
Hi! Thank you for this very detailed analysis.

This was a blind spot and I learned something fascinating. I have attempted to fix this via [this commit](https://github.com/GeoArkadeep/Stresslog/commit/ad3190a271070303b8cf2b387a1ab9f805348f3c).

I must mention that the JOSS review process is every bit as enjoyable and enriching for me as has been described. 🙂
