### 2025-08-28 GeoArkadeep/Stresslog#10 @frank1010111
Yeah, it's a surprising behavior for Python. But your solution looks great!
