### 2025-08-30 e8e9ec2
Update license and docstring in datasets_to_las
Replaced the previous copyright and license notice with the MIT License and updated the docstring format for datasets_to_las to use NumPy-style documentation.
