+++ stresslog/thirdparty.py
@@ -5,21 +5,41 @@ import pandas as pd
 
 def datasets_to_las(path, datasets, custom_units=None, **kwargs):
     """
-    copyright 2021 Agile Scientific
-    license
-    Apache 2.0
-    """
-    """
+    MIT License
+
+    Copyright (c) 2013-2020 Kent Inverarity and contributors
+
+    Permission is hereby granted, free of charge, to any person obtaining a copy
+    of this software and associated documentation files (the "Software"), to deal
+    in the Software without restriction, including without limitation the rights
+    to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+    copies of the Software, and to permit persons to whom the Software is
+    furnished to do so, subject to the following conditions:
+
+    The above copyright notice and this permission notice shall be included in all
+    copies or substantial portions of the Software.
+
+    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+    IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+    AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+    OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+    SOFTWARE.
+
     Write datasets to a LAS file on disk.
 
-    Args:
-        path (Str): Path to write LAS file to. If None, returns string buffer.
-        datasets (Dict['<name>': pd.DataFrame]): Dictionary maps a
-            dataset name (e.g. 'Curves') or 'Header' to a pd.DataFrame.
-        custom_units (Dict[str, str], optional): Dictionary mapping curve names to their units.
-            If a curve's unit is not specified, it defaults to an empty string.
-    Returns:
-        str if path is None, otherwise writes to file
+    Parameters
+    ----------
+    path : Str
+        Path to write LAS file to. If None, returns string buffer.
+    datasets : Dict['<name>': pd.DataFrame]
+        Dictionary maps a dataset name (e.g. 'Curves') or 'Header' to a pd.DataFrame.
+    custom_units : Dict[str, str], optional
+        Dictionary mapping curve names to their units. If a curve's unit is not specified, it defaults to an empty string.
+    Returns
+    -------
+    str if path is None, otherwise writes to file
     """
     from functools import reduce
     import warnings
