### 2025-08-30 320dffc
Clarify return value in datasets_to_las docstring
Updated the docstring for datasets_to_las to specify that the function returns a string if path is None, otherwise writes to file and returns None. This improves clarity for users of the function.
