### 2025-08-30 a5446fe
Clarify return value in datasets_to_las docstring
Updated the docstring to specify that if path is None, the function returns the las file as a string. This improves clarity for users of the function.
