### 2025-08-30 e829908
Clarify return values in datasets_to_las docstring
Updated the docstring for datasets_to_las to more clearly describe the function's return values depending on the 'path' argument.
