### 2025-08-30 5d492a4
Fix indentation in docstring for draw function
Corrected the indentation of the docstring list items in the draw() function to improve readability and maintain consistent formatting.
