### 2025-08-30 947342b
Fix indentation in docstring for drawSP function
Corrected the indentation of the docstring bullet points in the drawSP function to improve readability and maintain consistent formatting.
