+++ stresslog/DrawSP.py
@@ -73,10 +73,10 @@ def drawSP(Sv,Pp,bhp,shmin,UCS = 0,phi = 0, flag = 0,mu = 0.65,nu=0.25,bt=0,ym=0
     (for example Sv, Pp, bhp and ucs can all be in psi and the plot will
     also be in psi)
     This function creates a stress polygon visualization that includes:
-    - Stress regime domains (Normal, Strike-slip, Reverse)
-    - Breakout analysis with UCS variations
-    - Drilling-induced tensile fracture (DITF) analysis
-    - Stress bounds and interpolated values
+        - Stress regime domains (Normal, Strike-slip, Reverse)
+        - Breakout analysis with UCS variations
+        - Drilling-induced tensile fracture (DITF) analysis
+        - Stress bounds and interpolated values
     """
     
     #mu = (1-(2*nu))/(2*((nu*(1-nu))**0.5))
