### 2025-08-30 3e1967c
Add spacing in drawSP function docstring
Inserted a blank line in the drawSP function docstring for improved readability and separation of documentation sections.
