### 2025-08-30 828d8e0
Add docstring to find_TVD and minor doc update
Added a detailed docstring to the find_TVD function in geomechanics.py for improved clarity. Minor punctuation update to the drawSP function docstring in DrawSP.py.
