### 2025-08-30 b84a0c9
Fix indentation in getColumnHeights docstring
Corrected the indentation of the return value description in the getColumnHeights function docstring for improved readability and consistency.
