### 2025-08-30 bb91a5c
Fix indentation in getHoop docstring
Corrected the indentation of list items in the getHoop function's docstring for improved readability and consistency.
