### 2025-08-30 a62222b
Improve docstring formatting in BoreStab.py and DrawSP.py
Refined docstring formatting for better readability and consistency. Updated indentation and bullet points in parameter and return sections to enhance clarity for users and maintainers.
