### 2025-08-30 cd1d250
Update stress calculation functions and docstrings
Marked get_optimal_old as deprecated in BoreStab.py and updated its docstring to recommend get_principal_stress. Improved indentation in getVertical docstring. In obgppshmin.py, removed unused parameter wdf from get_OBG_pascals_vec, updated docstring formatting, and corrected offshore water pressure calculation.
