### 2025-08-30 5dcb383
Fix docstring formatting in getVertical function
Corrected line break in the docstring for getVertical to improve readability and maintain consistent formatting.
