### 2025-08-30 717453d
Update docstrings to NumPy style in fracture gradient functions
Revised the docstrings for get_Shmin_grad_Daine_ppg and get_Shmin_grad_Daine_ppg_vec to use NumPy documentation style for improved clarity and consistency.
