### 2025-08-30 c641ba4
Clarify input requirements in vectorized functions
Added notes to docstrings specifying that inputs may be scalars or 1D arrays and that all array inputs must have the same length. This clarifies broadcasting behavior and error handling for mismatched input lengths in vectorized calculations.
