### 2025-08-30 1e89466
Fix docstring indentation and example formatting
Corrected indentation in the get_analog function docstring and updated the example format for formation tops to improve clarity and consistency.
