### 2025-08-30 bdfe933
Fix docstring indentation in get_analog function
Corrected indentation in the docstring for the 'debug' parameter and return value section of the get_analog function to improve readability and consistency.
