### 2025-08-30 0c15381
Update forecast.py
