### 2025-08-30 daef5a8
Improve docstring for get_principal_stress function
Expanded and clarified the docstring for get_principal_stress to use NumPy-style formatting, provide detailed parameter descriptions, return values, and notes about the function's behavior and assumptions.
