### 2025-08-30 b8ab672
Update docstring for get_principal_stress function
Clarified the description of the 'dalpha' parameter in the docstring to indicate its default value is 0, removing internal implementation details.
