### 2025-08-30 a546f80
Improve docstring for plot_logs_labels function
Expanded and clarified the docstring for the plot_logs_labels function, providing detailed parameter descriptions, usage examples, and return value documentation to improve code readability and maintainability.
