### 2025-08-30 e3aeab3
Add spacing to docstrings in plotfrac function
Added blank lines to the docstring sections in the plotfrac function to improve readability and clarify parameter and return value documentation.
