### 2025-08-30 433bd91
Remove redundant notes from plotfrac docstring
Cleaned up the plotfrac function docstring by removing unnecessary notes about bit diameter and special case handling, making the documentation more concise.
