### 2025-08-30 14a41fd
Clarify docstring for plotfrac return values
Updated the docstring in plotfrac to better describe the meaning of cdepths and fdepths, specifying their relation to fracture regions and trends.
