### 2025-08-30 233061d
Improve docstrings for curve management functions
Expanded and clarified docstrings for add_curves and remove_curves functions in geomechanics.py. The updates provide more detailed parameter descriptions, return values, and notes for better usability and understanding.
