### 2025-08-30 a5bdeb0
Update docstring formatting in add_curves function
Reformatted the Notes section in the add_curves function docstring to remove unnecessary list formatting and improve readability.
