### 2025-08-30 6e38011
Bump version to 1.6.17
Updated project version to 1.6.17 in conf.py, pyproject.toml, and setup.py for a new release.
