### 2025-08-30 GeoArkadeep/Stresslog#9 @GeoArkadeep
Thank you again @frank1010111 for taking the time to prepare this incredibly detailed checklist. The api docs look much better with these implemented.

I have updated all of the mentioned docstrings via [these commits](https://github.com/GeoArkadeep/Stresslog/compare/ad3190a271070303b8cf2b387a1ab9f805348f3c...a5bdeb0759f1da6ea2a2cd9c7b5de17e8ddf66fc).
