### 2025-09-03 GeoArkadeep/Stresslog#9 @frank1010111
Looks great!
