### 2025-09-08 GeoArkadeep/Stresslog#11 @frank1010111
Test coverage
Running a test coverage report leads to these numbers:

```
Name                            Stmts   Miss  Cover
---------------------------------------------------
stresslog/BoreStab.py             414     72    83%
stresslog/DrawSP.py               244     49    80%
stresslog/PlotVec.py               54      2    96%
stresslog/Plotter.py              274     98    64%
stresslog/__init__.py              14      0   100%
stresslog/__main__.py               0      0   100%
stresslog/failure_criteria.py     111     13    88%
stresslog/forecast.py             205     86    58%
stresslog/geomechanics.py        1926    332    83%
stresslog/hydraulics.py           100     30    70%
stresslog/obgppshmin.py            96     22    77%
stresslog/plotangle.py            127     15    88%
stresslog/syntheticLogs.py        402    231    43%
stresslog/thirdparty.py            60      5    92%
stresslog/unit_converter.py        83     45    46%
---------------------------------------------------
TOTAL                            4110   1000    76%
```

Are there plans to add tests for the three files that have below 60% test coverage?

This is part of [JOSS review 8-36](https://github.com/openjournals/joss-reviews/issues/8036).
