+++ tests/test_units_stresslog.py
@@ -9,21 +9,21 @@ ATOL = 0.051  # Absolute Tolerance for 1-decimal rounding
 # --- Phase 1: Scalar Function Regression Tests ---
 
 @pytest.mark.parametrize("inputs, expected_output", [
-    ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
- 6000.0]), 'porepressures': np.array([3200.0, 3464.8, 6070.9, 8790.7, 10603.2, 12037.5, 13928.6, 15543.7,
- 18926.5, 21039.8])}, 2.3),
-    ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
- 6000.0]), 'porepressures': np.array([1108.1, 2316.3, 6557.0, 6792.8, 11063.0, 12879.0, 15308.5, 17338.0,
- 16869.5, 20910.7])}, 2.2),
-    ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
- 6000.0]), 'porepressures': np.array([2320.9, 2439.8, 5828.6, 8851.9, 9860.2, 11649.9, 11761.3, 16160.0,
- 18580.8, 20585.7])}, 2.2),
-    ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
- 6000.0]), 'porepressures': np.array([612.8, 3622.1, 5396.3, 6268.1, 9224.5, 12296.7, 14860.8, 15723.8, 18909.8,
- 20765.6])}, 2.2),
-    ({'tvds': np.array([1000.0, 1555.6, 2111.1, 2666.7, 3222.2, 3777.8, 4333.3, 4888.9, 5444.4,
- 6000.0]), 'porepressures': np.array([2700.1, 3690.9, 6820.5, 8388.6, 10468.3, 10381.0, 12522.3, 14158.6,
- 17927.8, 19984.7])}, 2.1),
+    ({'tvds': np.array([1000.0000, 1555.5556, 2111.1111, 2666.6667, 3222.2222, 3777.7778,
+ 4333.3333, 4888.8889, 5444.4444, 6000.0000]), 'porepressures': np.array([605.7176, 2552.0289, 4674.7912, 8224.7070, 9658.6526, 11795.4031,
+ 14345.6867, 16588.5969, 16450.7648, 19613.6491])}, 2.0973),
+    ({'tvds': np.array([1000.0000, 1555.5556, 2111.1111, 2666.6667, 3222.2222, 3777.7778,
+ 4333.3333, 4888.8889, 5444.4444, 6000.0000]), 'porepressures': np.array([581.2288, 3177.5257, 5688.3837, 7422.5514, 9579.3750, 10764.5609,
+ 14923.7025, 15970.1787, 18425.2006, 21308.3180])}, 2.1641),
+    ({'tvds': np.array([1000.0000, 1555.5556, 2111.1111, 2666.6667, 3222.2222, 3777.7778,
+ 4333.3333, 4888.8889, 5444.4444, 6000.0000]), 'porepressures': np.array([2457.0878, 3150.8340, 5718.8465, 8260.3096, 9370.9079, 13017.7722,
+ 12673.3184, 15860.9319, 18298.2248, 20951.6791])}, 2.2026),
[…]
