### 2025-09-10 e647d0a
Fix warnings about divide by zero
