+++ stresslog/geomechanics.py
@@ -1111,7 +1111,7 @@ def compute_geomech(
     else:
         shaleflag = np.zeros(len(md))
         zoneflag = np.zeros(len(md))
-    shaleflagN = (np.max(shaleflag) - shaleflag[:]) / np.max(shaleflag)
+    shaleflagN = np.divide(np.max(shaleflag) - shaleflag,np.max(shaleflag),out=np.full_like(shaleflag, np.nan, dtype=float),where=np.max(shaleflag) != 0)
     flag = Curve(shaleflag, mnemonic='ShaleFlag', units='ohm/m', index=md,
         null=0)
     zone = Curve(zoneflag, mnemonic='ShaleFlag', units='ohm/m', index=md,
@@ -1560,7 +1560,7 @@ def compute_geomech(
     else:
         shaleflag = np.zeros(len(md))
         zoneflag = np.zeros(len(md))
-    shaleflagN = (np.max(shaleflag) - shaleflag[:]) / np.max(shaleflag)
+    shaleflagN = np.divide(np.max(shaleflag) - shaleflag,np.max(shaleflag),out=np.full_like(shaleflag, np.nan, dtype=float),where=np.max(shaleflag) != 0)
     flag = Curve(shaleflag, mnemonic='ShaleFlag', units='ohm/m', index=md,
         null=0)
     zone = Curve(zoneflag, mnemonic='ShaleFlag', units='ohm/m', index=md,
@@ -1617,7 +1617,7 @@ def compute_geomech(
     hydroppf = 0.4335275040012 * hydrostatic
     mudppf = 0.4335275040012 * mudhydrostatic
     lithostatic = 2.6 * 9.80665 / 6.89476 * tvd
-    gradient = lithostatic / tvdf * 1.48816
+    gradient = np.divide(lithostatic, tvdf, out=np.full_like(lithostatic, np.nan), where=tvdf!=0) * 1.48816
     rhoppg[np.nanargmin(abs(tvdbgl))] = rhoappg
     if np.nanargmin(abs(tvdbgl)) > 0:
         rhoppg[0:np.nanargmin(abs(tvdbgl))] = np.nan
@@ -1771,8 +1771,7 @@ def compute_geomech(
         mvindex = np.nan
     else:
         mvindex = find_nearest_depth(tvd, maxveldepth)[0]
-    deltmu0 = np.nanmean(dalm[find_nearest_depth(tvd, maxveldepth)[0] - 5:
-        find_nearest_depth(tvd, maxveldepth)[0] + 5])
+    deltmu0 = np.nanmean(dalm[find_nearest_depth(tvd, maxveldepth)[0] - 5: find_nearest_depth(tvd, maxveldepth)[0] + 5]) if ul_depth>0 else np.nan
     c = ct
     b = ct
     print('Max velocity is ', deltmu0, 'uspf') if debug else None
@@ -2494,12 +2493,12 @@ def compute_geomech(
     skip = 21 if 2.0 <= window < 21 else window
     mtol = 1 - 2 * mabw / 360
     for i in range(0, len(tvd), 1):
-        sigmaVmpa = np.nanmean(obgpsi[i]) / 145.038
-        sigmahminmpa = np.nanmean(psifg[i]) / 145.038
-        sigmaHMaxmpa = np.nanmean(sgHMpsi[i]) / 145.038
-        ppmpa = np.nanmean(psipp[i]) / 145.038
-        bhpmpa = np.nanmean(mudpsi[i]) / 145.038
-        ucsmpa = np.nanmean(horsrud[i])
+        sigmaVmpa   = obgpsi[i]   / 145.038 if not np.isnan(obgpsi[i])   else np.nan
[…]
