### 2025-09-10 e63157d
improve test coverage
