+++ stresslog/forecast.py
@@ -466,63 +466,3 @@ def get_analog(well,current_forms, target_forms, kb,gl,dev=None,kop=0,ma=0,rob=0
     nfd["Start"] = offset1
     
     return get_analog_well(well,cfd,nfd,5,blankwell=newblankwell,debug=debug)
-    
-
-
-if __name__=="__main__":
-    import welly
-    spacing = 5
-    wellu = welly.Well.from_las("testlas.las")
-    print(wellu)
-    dev = pd.read_csv("testdev.csv")
-    print(dev)
-
-    well = getwelldev(wella=wellu,deva=dev,step=spacing)
-    print(well.df())
-    
-    well.header = pd.concat([well.header[well.header["mnemonic"] != "KB"], pd.DataFrame({"mnemonic": ["KB"], "value": [80]})], ignore_index=True)
-    well.header = pd.concat([well.header[well.header["mnemonic"] != "GL"], pd.DataFrame({"mnemonic": ["GL"], "value": [20]})], ignore_index=True)
-    
-    #well = getwelldev(wella=create_random_well(kb=35, gl=-200, step=spacing),step=spacing, kickoffpoint=2500, final_angle=25, rateofbuild=0.1, azimuth=270)
-    form_md = """TopTVD,StratiNum,Name,GR_Cut,Struc_Top,Struc_Bot,Z_ratio,OWC,GOC,Bt,Shm_Azim,Dip_Azim,Dip_Angle
-    1690,1,Alpha,70,500,2380,0.5,0,0,,,,
-    2380,2,Beta,70,1600,2610,0.5,0,0,,,,
-    2900,3,Gamma,70,2450,2950,0.5,2499,0,,90,,
-    2950,4,Echo,60,2550,3400,0.66,3100,3050,,90,0,0
-    3250,5,Delta,65,2750,3300,0.33,2950,2900,,85,2,3"""
-    
-    f0 = pd.read_csv("testforms.csv")
-    print(f0)
-
-    #f0 = pd.read_csv(io.StringIO(form_md), sep=",")
-    cfd = extract_depth_dict(convert_df_tvd(f0,well))
-    
-    new_md = """topTVD,StratiNum,Formation_Name,GR_Cut,Struc_Top,Struc_Bot,Z_ratio,OWC,GOC,Bt,Shm_Azim,Dip_Azim,Dip_Angle
-    1750,1,Alpha,70,500,2380,0.5,0,0,,,,
-    2124,2,Beta,70,1600,2610,0.5,0,0,,,,
-    2900,3,Gamma,70,2450,2950,0.5,2499,0,,90,,
-    3150,4,Echo,60,2550,3400,0.66,3100,3050,,90,0,0
-    4200,5,Delta,65,2750,3300,0.33,2950,2900,,85,2,3"""
-    
-    f1 = pd.read_csv("testforms2.csv")
-    print(f1)
-    
-    #f1 = pd.read_csv(io.StringIO(new_md), sep=",")
-    #newblankwell=create_blank_well(kb=20,gl=10,spacing=15,maxangle=90,kop=4000,azimuth=90)
-    
-    new_well,stringlas,devdas = get_analog(well,f0,f1,kb=20,gl=-200,ma=90,kop=1000,azi=90,debug=False)
-    freshnew = getwelldev(string_las=stringlas.read(),deva=devdas,step=5)
-    print(freshnew)
-    print(new_well)
-
-    output = compute_geomech(freshnew,sfs=30, tango=4000, writeFile=False,)
-    outdf = output[0]
-    outdf.to_csv("testout.csv")
-    print(new_well.df())
-    stringlas.seek(0)
-    print(stringlas.read(2500))
-    """
-    lulu = io.StringIO(output[1])
-    lulu.seek(0)
-    print(lulu.read(2500))
-    """
