### 2025-09-10 c551026
pragmatic solution to dlis tests
