+++ tests/test_units_stresslog.py
@@ -189,11 +189,11 @@ def test_getSP(inputs, expected_output):
 
 
 @pytest.mark.parametrize("inputs, expected_output", [
-    ({'s1': 144.0360, 's2': 29.2360, 's3': 46.0110, 'alpha': 113.1556, 'beta': 16.1858, 'gamma': 103.5780, 'azim': -73.7462, 'inc': -39.1889, 'theta': 175.3328, 'deltaP': 19.8546, 'Pp': 11.2787, 'nu': 0.2627}, (-20.3820, -45.2208, -21.3581, -8.0949, -57.5078, -60.0887, [-1.7980, 91.9749, 88.1922, 89.7175, 178.0050, 9.9449])),
-    ({'s1': 114.3041, 's2': 119.0893, 's3': 51.3233, 'alpha': -52.6641, 'beta': -16.3785, 'gamma': 136.7691, 'azim': -112.0377, 'inc': 8.4108, 'theta': 166.5956, 'deltaP': 107.9572, 'Pp': 120.0245, 'nu': 0.3919}, (-77.2157, 33.0977, -1.8043, 33.1272, -77.2453, -0.9368, [-158.8964, 177.4465, 0.2861, 92.3870, 90.9065, 89.6761])),
-    ({'s1': 176.2960, 's2': 126.2873, 's3': 21.5282, 'alpha': -152.3213, 'beta': -64.5810, 'gamma': 142.8582, 'azim': -53.2580, 'inc': 153.0492, 'theta': -155.4493, 'deltaP': 63.8631, 'Pp': 25.7986, 'nu': 0.4199}, (-91.6976, 92.4353, 17.0107, 93.9937, -93.2559, 5.2341, [87.5590, 87.3420, 2.4933, 91.1272, 177.1126, 64.5468])),
-    ({'s1': 159.8108, 's2': 65.0776, 's3': 22.2195, 'alpha': 92.2066, 'beta': -14.9570, 'gamma': -54.4390, 'azim': 124.4123, 'inc': 19.2097, 'theta': -143.0936, 'deltaP': 36.7492, 'Pp': 88.3991, 'nu': 0.3330}, (77.8673, 15.2701, -109.9233, 160.8610, -67.7236, -52.9467, [89.0572, 91.6093, -0.9501, 90.2610, 178.3697, -100.1615])),
-    ({'s1': 80.9919, 's2': 114.6084, 's3': 11.2391, 'alpha': 49.9817, 'beta': -101.8218, 'gamma': -1.5150, 'azim': -54.6688, 'inc': 72.1373, 'theta': 109.4874, 'deltaP': 133.8510, 'Pp': 47.6181, 'nu': 0.4012}, (12.5561, 155.3661, 23.9920, 159.2890, 8.6332, 9.2861, [89.9597, 179.1280, 1.7771, 90.0264, 89.1284, 88.2225])),
+    ({'s1': 35.4274, 's2': 84.4982, 's3': 11.7613, 'alpha': -175.7341, 'beta': 103.0121, 'gamma': 2.5293, 'azim': -165.2040, 'inc': -152.3231, 'theta': -84.6083, 'deltaP': 128.5445, 'Pp': 124.4367, 'nu': 0.2113}, (-417.9725, -145.8275, 9.9480, -145.4643, -418.3357, 2.0907, [87.3791, 176.9312, 1.7979, 90.0441, 93.0685, -88.2045])),
[…]
