### 2025-09-12 GeoArkadeep/Stresslog#11 @GeoArkadeep
I have added some unit tests to increase coverage in unit converter. Forecast had it's standalone test hook section removed to bring it above 60 percent. sytntheticLogs.py had a lot of code dealing with dlis files, which now has pragma no cover (because we cannot test the dlis functionality without an actual dlis file, and I dont want to mock dlisio). Overall coverage is now at 80 percent and all are above 60.
