### 2025-09-13 GeoArkadeep/Stresslog#11 @frank1010111
Great! In the future I'd recommend adding a dlis file for testing, but this is a good improvement.
