### 2025-09-15 #8036 @willgearty
Great to hear, thanks for being so thorough @frank1010111! And thanks for your acceptance as well @wkearn! Both of you have really gone above and beyond to help improve this piece of software and make it suitable for JOSS!
