### 2025-09-15 36af8ea
Bump version to 1.7.1
Update version number to 1.7.1 in conf.py, pyproject.toml, and setup.py to prepare for a new release, to test out Zenodo balling
