### 2025-09-16 9633b2a
Fix typos and capitalization
