### 2025-09-16 f412d8d
Bump version
Final bump before JOSS publication
