### 2025-09-16 #8036 @GeoArkadeep
I have bumped the version to 1.7.2
The new Zenodo doi is https://doi.org/10.5281/zenodo.17128441
