### 2025-09-16 #8036 @GeoArkadeep
So sorry about that.

Version is now 1.7.4
Zenodo doi: https://doi.org/10.5281/zenodo.17132819
