### 2025-09-16 #8036 @willgearty
Thanks for bearing with me @GeoArkadeep! Just need to get the track editor to look everything over.
