### 2025-09-17 #8036 @willgearty
@kthyng, is this an issue? The length of the paper was a topic of discussion during review, but we decided the extra text was important.
