### 2025-09-19 #8036 @kthyng
@willgearty @GeoArkadeep It looks like an item in the paper is too long, not the paper itself. Looking through, the biggest thing was the equation. Could this be referenced from the docs instead? Alternatively, does something else look like it could be triggering the error?
