### 2025-09-19 d5be470
attempt to reduce font size attempt 2
