### 2025-09-19 691ee32
Bump version for JOSS release
