### 2025-09-19 #8036 @GeoArkadeep
@proceedings{10.2118/107785-MS,
    author = {Zhang, J. and Standifird, W. B. and Shen, X.},
    title = {Optimized Perforation Tunnel Geometry, Density and Orientation to Control Sand Production},
    volume = {European Formation Damage Conference},
    series = {SPE European Formation Damage Conference and Exhibition},
    pages = {SPE-107785-MS},
    year = {2007},
    month = {05},
    doi = {10.2118/107785-MS},
    url = {https://doi.org/10.2118/107785-MS},
}

This seems to be the offending citation, indeed volume is 36 chars long.

May I set it to volume_title instead?
