### 2025-09-19 0c2c15c
Changed citation slightly, bump version
