### 2025-09-19 #8036 @willgearty
I think `volume_title` is used in the XML when there is a volume and the title of the needs to have a a colon (:) afterwards for citation style purposes. I don't think `volume_title` is a valid bibtex field, based on the [docs](https://bibtex.eu/fields/). Since the volume is just repeating the series name, I would probably recommend removing the `volume` field for this reference.
