+++ paper/paper.bib
@@ -137,7 +137,6 @@
 @proceedings{Zhang2007,
     author = {Zhang, J. and Standifird, W. B. and Shen, X.},
     title = {Optimized Perforation Tunnel Geometry, Density and Orientation to Control Sand Production},
-    volume_title = {European Formation Damage Conference},
     series = {SPE European Formation Damage Conference and Exhibition},
     pages = {SPE-107785-MS},
     year = {2007},
