### 2025-09-19 #8036 @GeoArkadeep
Zenodo archive: https://doi.org/10.5281/zenodo.17162503
