### 2025-09-20 #8036 @kthyng
@GeoArkadeep We recommend that the title and author list for the metadata in your Zenodo software archive exactly match those in your JOSS paper to create a cohesive archive, though this is not a strict requirement. Do you want to make this change?
