### 2025-09-25 #8036 @kthyng
@GeoArkadeep
* Please check that your inline vs. parenthetical references are correct. You can compare with the example paper here: https://joss.readthedocs.io/en/latest/example_paper.html. For example, in your summary the reference should be inline not parenthetical, also later in the paper.
* when referencing packages, especially those that aren't well known, please use `` around the package name so people can distinguish the names. Check in particular the last paragraph of the Statement of Need. Similarly, the "las" and "dlis" files could be better distinguished somehow.
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
* Should "python" be capitalized throughout your paper?
