### 2025-09-26 07f980c
Clarify file extensions and library names in documentation
Updated references to file types and libraries in the documentation for consistency and clarity, specifying '.dlis' and '.las' file extensions and formatting library names with backticks.
