### 2025-09-26 552bc57
Update library names for consistency in paper
Corrected capitalization and naming of libraries (NumPy, pandas, SciPy, plotly.py) to ensure consistency and accuracy in documentation.
