### 2025-09-26 6c1d294
inline vs parenthetical citation correction
Replaced bracketed citation syntax ([@pevska1995], [@Thompson2022], [@northernlights]) with plain references (e.g., @pevska1995) throughout the document for consistency and improved formatting.
