+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 # Summary
 
-This package is meant to be used by researchers and practitioners working in the field of geomechanics. It uses a collection of algorithms used to iteratively model the sate of stress underground, given a well log. The computations use 6 component stress tensor, as calculated using [@pevska1995], which allows modeling of inclined wellbores in inclined states of stress.
+This package is meant to be used by researchers and practitioners working in the field of geomechanics. It uses a collection of algorithms used to iteratively model the sate of stress underground, given a well log. The computations use 6 component stress tensor, as calculated using @pevska1995, which allows modeling of inclined wellbores in inclined states of stress.
 
 # Statement of need
 
@@ -55,7 +55,7 @@ $$
 
 Where $R_s$ is the rotation matrix defined by Euler angles $\alpha$, $\beta$ and $\gamma$, in the NED reference frame.
 
-Considering the azimuth of maximum principal stress as $\alpha$, the above relations are used to optimize for the angles $\beta$ and $\gamma$. The technique proposed by [@pevska1995] starts with good estimates of the far field principal stresses, $\sigma_1$, $\sigma_2$ and $\sigma_3$, already rotated by the Euler Angles $\alpha$, $\beta$ and $\gamma$. Usually, however, what is available is an estimate of minimum horizontal stress, an estimate of the vertical stress, and an estimate of maximum horizontal stress. It is insufficient to simply rotate the tensor, as the rotated tensor will not have the correct vertical component. We optimize the principal stresses ($\sigma_1$, $\sigma_2$ and $\sigma_3$) such that the vertical and horizontal components of the tensor match the specified vertical and horizontal stresses.
+Considering the azimuth of maximum principal stress as $\alpha$, the above relations are used to optimize for the angles $\beta$ and $\gamma$. The technique proposed by @pevska1995 starts with good estimates of the far field principal stresses, $\sigma_1$, $\sigma_2$ and $\sigma_3$, already rotated by the Euler Angles $\alpha$, $\beta$ and $\gamma$. Usually, however, what is available is an estimate of minimum horizontal stress, an estimate of the vertical stress, and an estimate of maximum horizontal stress. It is insufficient to simply rotate the tensor, as the rotated tensor will not have the correct vertical component. We optimize the principal stresses ($\sigma_1$, $\sigma_2$ and $\sigma_3$) such that the vertical and horizontal components of the tensor match the specified vertical and horizontal stresses.
 
 For every depth-sample, the stresses resolved on the wellbore wall are calculated along the circumference. The lower critical mudweight is calculated by using the modified Lade formula [@ewy1999] for critical mudweight during this process. A closed-form solution has been derived by setting $\sigma_{\theta_{\min}}$ equal to tensile stress and solving this for the upper critical mud pressure, as follows:
 
@@ -122,13 +122,13 @@ From observations on multiple wells sampling the same stress field at different 
 
 # Case Study
 
-The well data from Equinor Northern Lights dataset [@northernlights] has been used as the example here, to model the stress state occurring in the depth interval of 2600 to 2630m. The resistivity image log shows the occurrence of en-echelon fractures in a vertical wellbore. The model applied here uses parameters similar to [@Thompson2022], with a stress tensor tilt of 2 degrees towards south, and is able to replicate the fracture patterns observed in the actual image log.
[…]
