### 2025-09-26 989187d
preserve case for references
Updated all BibTeX entry titles, review the citation for Traugott 97 and if possible add one for Pint
