+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Thompson2022,
-    title = {Characterization of the in-situ stress on the Horda platform – A study from the Northern Lights Eos well},
+    title = {{Characterization of the in-situ stress on the Horda platform – A study from the Northern Lights Eos well}},
     journal = {International Journal of Greenhouse Gas Control},
     volume = {114},
     pages = {103580},
@@ -12,7 +12,7 @@
 
 @inbook{Flemings2021,
     place={Cambridge},
-    title={Flow Focusing and Centroid Prediction},
+    title={{Flow Focusing and Centroid Prediction}},
     booktitle={A Concise Guide to Geopressure: Origin, Prediction, and Applications},
     publisher={Cambridge University Press},
     author={Flemings, Peter B.},
@@ -22,7 +22,7 @@
 }
 
 @article{Zhang20132,
-    title = {Effective stress, porosity, velocity and abnormal pore pressure prediction accounting for compaction disequilibrium and unloading},
+    title = {{Effective stress, porosity, velocity and abnormal pore pressure prediction accounting for compaction disequilibrium and unloading}},
     journal = {Marine and Petroleum Geology},
     volume = {45},
     pages = {2-11},
@@ -34,7 +34,7 @@
 }
 
 @article{pevska1995,
-    title={Compressive and tensile failure of inclined well bores and determination of in situ stress and rock strength},
+    title={{Compressive and tensile failure of inclined well bores and determination of in situ stress and rock strength}},
     author={Pe{\v{s}}ka, Pavel and Zoback, Mark D},
     journal={Journal of Geophysical Research: Solid Earth},
     volume={100},
@@ -47,7 +47,7 @@
 
 @proceedings{willson2002,
     author = {Willson, S. M. and Moschovidis, Z. A. and Cameron, J. R. and Palmer, I. D.},
-    title = {New Model for Predicting the Rate of Sand Production},
+    title = {{New Model for Predicting the Rate of Sand Production}},
     volume = {All Days},
     series = {SPE/ISRM Rock Mechanics Conference},
     pages = {SPE-78168-MS},
@@ -59,7 +59,7 @@
 }
 
 @article{horsrud2001,
-    title={Estimating mechanical properties of shale from empirical correlations},
+    title={{Estimating mechanical properties of shale from empirical correlations}},
     author={Horsrud, Per},
     journal={SPE Drilling \& Completion},
     volume={16},
@@ -71,7 +71,7 @@
 }
 
 @inproceedings{lal1999,
-    title={Shale stability: drilling fluid interaction and shale strength},
+    title={{Shale stability: drilling fluid interaction and shale strength}},
     author={Lal, Manohar},
     booktitle={SPE Asia Pacific Oil and Gas Conference and Exhibition},
     pages={SPE--54356},
@@ -81,7 +81,7 @@
 }
 
 @article{thorsen2011,
-    title = {In situ stress estimation using borehole failures — Even for inclined stress tensor},
+    title = {{In situ stress estimation using borehole failures — Even for inclined stress tensor}},
     journal = {Journal of Petroleum Science and Engineering},
     volume = {79},
     number = {3},
@@ -94,7 +94,7 @@
 }
 
 @article{ZOBACK2003,
-    title = {Determination of stress orientation and magnitude in deep wells},
+    title = {{Determination of stress orientation and magnitude in deep wells}},
     journal = {International Journal of Rock Mechanics and Mining Sciences},
     volume = {40},
     number = {7},
@@ -108,7 +108,7 @@
 }
 
 @article{traugott1997,
[…]
