### 2025-09-26 4f5b59f
Further check and update case in references
updated zoback92, SciPy20. Another pass to follow immediately
