+++ paper/paper.bib
@@ -147,7 +147,7 @@
 }
 
 @article{zoback1992,
-  title={{In situ stress measurements to 3.5 km depth in the Cajon Pass scientific research borehole: implications for the mechanics of crustal faulting}},
+  title={{In situ stress measurements to 3.5 km depth in the Cajon Pass Scientific Research Borehole: Implications for the mechanics of crustal faulting}},
   author={Zoback, Mark D and Healy, John H},
   journal={Journal of Geophysical Research: Solid Earth},
   volume={97},
@@ -196,8 +196,8 @@
             Harris, Charles R. and Archibald, Anne M. and
             Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
             {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
-  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
-            Computing in Python}},
+  title   = {{{SciPy} 1.0: fundamental algorithms for scientific
+            computing in Python}},
   journal = {Nature Methods},
   year    = {2020},
   volume  = {17},
