+++ paper/paper.bib
@@ -59,7 +59,7 @@
 }
 
 @article{horsrud2001,
-    title={{Estimating mechanical properties of shale from empirical correlations}},
+    title={{Estimating Mechanical Properties of Shale From Empirical Correlations}},
     author={Horsrud, Per},
     journal={SPE Drilling \& Completion},
     volume={16},
@@ -273,8 +273,7 @@ year = {2023}
 @inproceedings{mckinney2010data,
   author       = {McKinney, Wes},
   title        = {{Data Structures for Statistical Computing in Python}},
-  booktitle    = {Proceedings of the 9th Python in Science Conference},
-  editor       = {van der Walt, Stéfan and Millman, Jarrod},
+  booktitle    = {{Proceedings of the 9th Python in Science Conference}},
   year         = {2010},
   pages        = {56--61},
   doi          = {10.25080/Majora-92bf1922-00a},
@@ -284,7 +283,7 @@ year = {2023}
 
 @Article{Hunter:2007,
   Author    = {Hunter, J. D.},
-  Title     = {{Matplotlib: A 2D graphics environment}},
+  Title     = {{Matplotlib: A 2D Graphics Environment}},
   Journal   = {Computing in Science \& Engineering},
   Volume    = {9},
   Number    = {3},
