### 2025-09-26 9f6fbc3
Add Pint citation and reference
