+++ paper/paper.bib
@@ -295,3 +295,12 @@ year = {2023}
   doi       = {10.1109/MCSE.2007.55},
   year      = 2007
 }
+
+@misc{pint,
+  author = {Hernan Grecco},
+  title = {Pint: Operate and manipulate physical quantities in Python},
+  year = {2012},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/hgrecco/pint}
+}
