### 2025-09-26 525c3c0
bump version to 1.7.8
