### 2025-09-26 #8036 @GeoArkadeep
Thank you all very much.
@kthyng thank you from coordinating the whole process from pre-review to publication!
@willgearty thank you for your help along the whole way, from recruiting reviewers to debugging issues with the paper!
@frank1010111 and @wkearn thank you both for enormously improving the quality of the repository and for the awesome feedback.

It was a very enjoyable process! I hope to be able to contribute more papers here in the future!!
